[library][xl]: refactor datapub to be a library
