+++ package.json
@@ -3,6 +3,15 @@
   "version": "0.1.0",
   "private": false,
   "main": "dist/index.js",
+  "module": "dist/index.js",
+  "files": [
+    "dist",
+    "README.md"
+  ],
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/datopian/datapub"
+  },
   "dependencies": {
     "axios": "^0.19.2",
     "ckanClient": "git+https://github.com/datopian/ckan-client-js.git",
@@ -17,7 +26,7 @@
   },
   "scripts": {
     "start": "react-scripts start",
-    "build": "react-scripts build",
+    "build": "rm -rf dist && NODE_ENV=production babel src/lib --out-dir dist --copy-files --ignore __tests__,spec.js,test.js,__snapshots__",
     "dist": "rm -rf dist/ && mkdir dist && mkdir dist/css && NODE_ENV=production babel src/ -d dist/ --copy-files",
     "test": "react-scripts test",
     "test:watch": "npm run test -- --coverage --watchAll",
