+++ src/lib/components/Switcher/Switcher.css
