+++ src/lib/index.js
@@ -1,0 +1,10 @@
+import Choose from "./components/Choose";
+import InputFile from "./components/InputFile";
+import InputUrl from "./components/InputUrl";
+import Metadata from "./components/Metadata";
+import ProgressBar from "./components/ProgressBar";
+import Switcher from "./components/Switcher";
+import TableSchema from "./components/TableSchema";
+import Upload from "./components/Upload";
+
+export { Upload, Choose, InputFile, InputUrl, Metadata, ProgressBar, Switcher, TableSchema };
